Removed bug from FlatInfo entity.

+++ rentalmatesbackend/src/main/java/com/example/vivek/rentalmates/backend/entities/FlatInfo.java
@@ -25,7 +25,7 @@ public class FlatInfo {
         return latitude;
     }
 
-    public void setLatitude(Long latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
@@ -33,7 +33,7 @@ public class FlatInfo {
         return longitude;
     }
 
-    public void setLongitude(Long longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
